requirements.txt
ultralytics~=8.2.91
PyYAML~=6.0.2
mlflow~=2.16.0
torch~=2.4.1